# openapi.json
{
  "components": {
    "schemas": {
      "AccountRefDto": {
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "employeeNo": {
            "title": "Employeeno",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "employeeNo",
          "displayName"
        ],
        "title": "AccountRefDto",
        "type": "object"
      },
...
      "LeaderNodeDto": {
        "properties": {
          "account": {
            "$ref": "#/components/schemas/AccountRefDto"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/AccountRefDto"
            },
            "title": "Members",
            "type": "array"
          }
        },
        "required": [
          "account",
          "members"
        ],
        "title": "LeaderNodeDto",
        "type": "object"
      },
...
      "ManagerNodeDto": {
        "properties": {
          "account": {
            "$ref": "#/components/schemas/AccountRefDto"
          },
          "leaders": {
            "items": {
              "$ref": "#/components/schemas/LeaderNodeDto"
            },
            "title": "Leaders",
            "type": "array"
          }
        },
        "required": [
          "account",
          "leaders"
        ],
        "title": "ManagerNodeDto",
        "type": "object"
      },
...
      "OrgTreeResponseDto": {
        "properties": {
          "managers": {
            "items": {
              "$ref": "#/components/schemas/ManagerNodeDto"
            },
            "title": "Managers",
            "type": "array"
          }
        },
        "required": [
          "managers"
        ],
        "title": "OrgTreeResponseDto",
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Type"
          }
        },
        "required": [
          "title",
          "status",
          "requestId"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "SetSuperiorRequestDto": {
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "superiorId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Superiorid"
          }
        },
        "required": [
          "reason"
        ],
        "title": "SetSuperiorRequestDto",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "内部研发平台 Control Plane API。\n\n全局约定：JSON 一律 camelCase；ID 一律 string；错误统一 application/problem+json（RFC 9457）；\n分页 cursor 型 {items, nextCursor}；写并发 If-Match/ETag；变更命令 Idempotency-Key——\n后三者自 V0.2 首个真实接口起强制执行。",
    "title": "engineering-platform-control-plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/accounts/{accountId}/superior": {
      "put": {
        "operationId": "org_set_superior",
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "title": "Accountid",
              "type": "string"
            }
          },
          {
            "description": "Stable replay key: 8-128 ASCII letters, digits, dot, underscore, colon, or hyphen",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "description": "Stable replay key: 8-128 ASCII letters, digits, dot, underscore, colon, or hyphen",
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSuperiorRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not ready"
          }
        },
        "summary": "Organization Set Superior",
        "tags": [
          "organization"
        ]
      }
    },
...
    "/api/v1/admin/organization/tree": {
      "get": {
        "operationId": "org_tree",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgTreeResponseDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not ready"
          }
        },
        "summary": "Organization Tree",
        "tags": [
          "organization"
        ]
      }
    },
...
  }
}